Refactor SanitizeOptions function.
