+++ util/options.cc
@@ -6,8 +6,19 @@
 
 #include "leveldb/comparator.h"
 #include "leveldb/env.h"
+#include "leveldb/cache.h"
+#include "db/filename.h"
 
 namespace leveldb {
+
+const int kNumNonTableCacheFiles = 10;
+
+// Fix user-supplied options to be reasonable
+template <class T,class V>
+static void ClipToRange(T* ptr, V minvalue, V maxvalue) {
+  if (static_cast<V>(*ptr) > maxvalue) *ptr = maxvalue;
+  if (static_cast<V>(*ptr) < minvalue) *ptr = minvalue;
+}
 
 Options::Options()
     : comparator(BytewiseComparator()),
@@ -25,6 +36,36 @@ Options::Options()
       compression(kSnappyCompression),
       reuse_logs(false),
       filter_policy(NULL) {
+}
+
+Options Options::Sanitize(const std::string& dbname,
+                          const Comparator* icmp,
+                          const FilterPolicy* ipolicy) const {
+  Options result = *this;
+  result.comparator = icmp;
+  result.filter_policy = (filter_policy != NULL) ? ipolicy : NULL;
+  ClipToRange(&result.max_open_files,    64 + kNumNonTableCacheFiles, 50000);
+  ClipToRange(&result.write_buffer_size, 64<<10,                      1<<30);
+  ClipToRange(&result.max_file_size,     1<<20,                       1<<30);
+  ClipToRange(&result.block_size,        1<<10,                       4<<20);
+  if (result.info_log == NULL) {
+    // Open a log file in the same directory as the db
+    env->CreateDir(dbname);  // In case it does not exist
+    env->RenameFile(InfoLogFileName(dbname), OldInfoLogFileName(dbname));
+    Status s = env->NewLogger(InfoLogFileName(dbname), &result.info_log);
+    if (!s.ok()) {
+      // No place suitable for logging
+      result.info_log = NULL;
+    }
+  }
+  if (result.block_cache == NULL) {
+    result.block_cache = NewLRUCache(8 << 20);
+  }
+  return result;
+}
+
+int Options::GetTableCacheSize() const {
+  return max_open_files - kNumNonTableCacheFiles;
 }
 
 }  // namespace leveldb
